blockId: Refactor into types + schemas
