+++ src/block/blockId/typeSchemas.ts
@@ -1,0 +1,7 @@
+import { JSONSchemaType } from "ajv";
+
+import type { BlockId } from "./types";
+
+const blockIdTypeSchema: JSONSchemaType<BlockId> = { type: "string" };
+
+export { blockIdTypeSchema };
