+++ src/block/blockId/types.ts
@@ -1,0 +1,3 @@
+type BlockId = string;
+
+export type { BlockId };
